pyproject.toml
[tool.aerich]
tortoise_orm = "database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.2"
pre-commit = "^4.2.0"

[project]
name = "tonversego"
version = "0.1.0"
description = "GeoNFT game"
authors = [
    {name = "Ivan Ivanov",email = "kriventsev.i@tonversego.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.0,<0.117.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "geopy (>=2.4.1,<3.0.0)",
    "aerich (>=0.9.1,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
